move unsupported on mysql from common
